// src/models/GuildConfig.d.ts
import { Schema } from 'mongoose';
export declare const GuildConfig: import("mongoose").Model<{
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
}, {}, {}, {
    id: string;
}, import("mongoose").Document<unknown, {}, {
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
}, {
    id: string;
}, import("mongoose").DefaultSchemaOptions> & Omit<{
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
} & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, Schema<any, import("mongoose").Model<any, any, any, any, any, any, any>, {}, {}, {}, {}, import("mongoose").DefaultSchemaOptions, {
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
}, import("mongoose").Document<unknown, {}, {
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
}, {
    id: string;
}, import("mongoose").DefaultSchemaOptions> & Omit<{
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
} & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, unknown, {
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
} & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}>, {
    guildId: string;
    clockInChannelId?: string | null;
    clockOutChannelId?: string | null;
    breakChannelId?: string | null;
    reportsChannelId?: string | null;
    dashboardChannelId?: string | null;
    dashboardMessageId?: string | null;
    lateAlertsChannelId?: string | null;
} & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}>;
//# sourceMappingURL=GuildConfig.d.ts.map